add glossary, style select
fix translation
